Refacto create product with combinations test
